chore: remove externals
