+++ packages/cli/package.json
@@ -23,7 +23,7 @@
     }
   },
   "bin": {
-    "dotsecret": "dist/cli.js"
+    "dotsecret": "dist/cli/index.js"
   },
   "scripts": {
     "build": "tsup",
@@ -36,13 +36,22 @@
   },
   "dependencies": {
     "axios": "1.6.7",
+    "boxen": "7.1.1",
+    "chalk": "5.3.0",
+    "color-json": "3.0.5",
     "commander": "12.0.0",
+    "cosmiconfig": "9.0.0",
     "dayjs": "1.11.10",
     "dotenv": "16.4.5",
+    "execa": "8.0.1",
     "fast-json-stable-stringify": "2.1.0",
+    "find-up": "7.0.0",
+    "globby": "14.0.1",
     "js-yaml": "4.1.0",
     "lodash": "4.17.21",
+    "multimatch": "7.0.0",
     "nunjucks": "3.2.4",
+    "ora": "8.0.1",
     "prompts": "2.4.2",
     "table": "6.8.1",
     "zod": "3.22.4"
@@ -55,16 +64,8 @@
     "@types/nunjucks": "3.2.6",
     "@types/prompts": "2.4.9",
     "@types/semver": "7.5.7",
-    "boxen": "7.1.1",
-    "chalk": "5.3.0",
-    "color-json": "3.0.5",
     "eslint-config-peppy": "3.1.1",
     "eslint": "8.56.0",
-    "execa": "8.0.1",
-    "find-up": "7.0.0",
-    "globby": "14.0.1",
-    "multimatch": "7.0.0",
-    "ora": "8.0.1",
     "tsup": "8.0.2",
     "typescript": "5.3.3",
     "yalc": "1.0.0-pre.53"
