+++ packages/cli/src/index.ts
@@ -1,1 +1,1 @@
-// TODO: This file should expose the public API of the dotsecret package that can be used by other node packages and projects without having to use the CLI.
+export { getConfig } from "@/lib/config";
