--- packages/cli/src/libs/command/index.ts
@@ -1,1 +1,0 @@
-export * from "./command";
