--- packages/cli/src/libs/fs/index.ts
@@ -1,1 +1,0 @@
-export * from "./fs";
