--- packages/cli/src/libs/issue/index.ts
@@ -1,1 +1,0 @@
-export * from "./issue";
