--- packages/cli/src/libs/logger/index.ts
@@ -1,1 +1,0 @@
-export * from "./logger";
