--- packages/cli/src/libs/secret/index.ts
@@ -1,1 +1,0 @@
-export * from "./secret";
