--- packages/cli/src/libs/template/index.ts
@@ -1,1 +1,0 @@
-export * from "./template";
