--- packages/cli/src/vendors/.eslintrc.cjs
@@ -1,7 +1,0 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  rules: {
-    "import/no-extraneous-dependencies": ["off"],
-    "no-restricted-imports": ["off"],
-  },
-};
