--- packages/cli/src/vendors/boxen.ts
@@ -1,4 +1,0 @@
-import boxen from "boxen";
-
-export * from "boxen";
-export default boxen;
