--- packages/cli/src/vendors/chalk.ts
@@ -1,4 +1,0 @@
-import chalk from "chalk";
-
-export * from "chalk";
-export default chalk;
