--- packages/cli/src/vendors/color-json.ts
@@ -1,4 +1,0 @@
-import colorJson from "color-json";
-
-export * from "color-json";
-export default colorJson;
