--- packages/cli/src/vendors/execa.ts
@@ -1,4 +1,0 @@
-import * as execa from "execa";
-
-export * from "execa";
-export default execa;
