--- packages/cli/src/vendors/find-up.ts
@@ -1,4 +1,0 @@
-import * as findUp from "find-up";
-
-export * from "find-up";
-export default findUp;
