--- packages/cli/src/vendors/globby.ts
@@ -1,4 +1,0 @@
-import * as globby from "globby";
-
-export * from "globby";
-export default globby;
