--- packages/cli/src/vendors/multimatch.ts
@@ -1,4 +1,0 @@
-import multimatch from "multimatch";
-
-export * from "multimatch";
-export default multimatch;
