--- packages/cli/src/vendors/ora.ts
@@ -1,4 +1,0 @@
-import ora from "ora";
-
-export * from "ora";
-export default ora;
