+++ packages/cli/test/.secretrc
@@ -1,0 +1,3 @@
+{
+  "gitignore": true
+}
