+++ packages/cli/test/foo/bar/baz/qux.json.secret
@@ -1,0 +1,1 @@
+{}
