+++ packages/cli/test/foo/bar/the-ignored-folder/.secretignore
@@ -1,0 +1,1 @@
+**/*
