+++ packages/cli/tsconfig.json
@@ -18,5 +18,5 @@
     }
   },
   "include": ["**/*.ts"],
-  "exclude": ["node_modules", "**/*.test.ts", "dist"]
+  "exclude": ["node_modules", "dist"]
 }
