+++ v1.md
@@ -21,9 +21,17 @@ There is the description of the project commands, files and secrets managers tha
 
 ## Files
 
-- **.secret.config**: Contains the configuration for the project
-- **.secret.local**: [.gitignore] Contains the secrets pulled from the secrets manager that can be edited and pushed back
-- **.secret.previous**: [.gitignore] Contains the previous version of the .secret.local before a pull
+- **Configuration files**: Contains the configuration for the project
+  - .secretrc
+  - .secretrc.json
+  - .secretrc.yaml
+  - .secretrc.yml
+  - .secretrc.js
+  - .secretrc.ts
+  - .secretrc.mjs
+  - .secretrc.cjs
+- **.secret**: [.gitignore] Contains the secrets pulled from the secrets manager that can be edited and pushed back
+- **.secret.previous**: [.gitignore] Contains the previous version of the .secret before a pull
 - **.secret.me**: [.gitignore] Contains the user credentials to connect to the secrets manager
 - **.secretignore**: Contains the files and directories that should be ignored by dotsecret
 - **\*.secret**: Template files that can be rendered with dotsecret render (eg: config.json.secret -> config.json)
@@ -38,3 +46,21 @@ There is the description of the project commands, files and secrets managers tha
 - **Azure Key Vault**: [TBD] https://azure.microsoft.com/en-us/products/key-vault
 - **Google Cloud Secret Manager**: [TBD] https://cloud.google.com/security/products/secret-manager
 - **Vercel Project Environment Variables**: [TBD] https://vercel.com/docs/projects/environment-variables
+
+## Dictionnary
+
+- **Template file**: A file that contains secrets placeholders that can be rendered with the dotsecret render command. Eg: config.json.secret
+- **Render**: The process of generating the rendered files from the template files
+- **Rendered file**: The file that is generated from the template file after the dotsecret render command. Eg: config.json
+- **Secrets manager**: A service that allows to store and retrieve secrets securely. Eg: HashiCorp Vault, Doppler, Infisical
+- **Secrets**: Sensitive data that should be kept secret. Eg: API keys, passwords, tokens
+- **Audit**: A process that checks the project for security issues
+- **Fix**: A process that tries to fix the security issues found by the audit
+- **Session**: A period of time during which the user is logged in
+- **User**: A person that uses dotsecret
+- **Project**: A software project that uses dotsecret
+- **Configuration**: The settings of the project
+- **Local secrets**: The secrets that are pulled from the secrets manager and can be edited and pushed back
+- **Previous local secrets**: The secrets that were pulled from the secrets manager before the last pull
+- **Me secrets**: The user credentials to connect to the secrets manager
+- **Ignore file**: A file that contains the files and directories that should be ignored by dotsecret when it's rendering the template files
